_find('apath') should be an identity
It's not possible to use find-as-merge with input trees that happen to
still be strings, because _find is incorrectly failing to notice that
each one should simply be an identity operation.

I added a test that covers this case, although the rest of the test
suite seems to be in bad shape for unrelated reasons.